fix: update provider for jamfprotect-go-sdk v0.2.0 compatibility
Computer.MemorySize changed from *int64 to *float64 in the SDK. Update
the ComputerModel field, schema attribute, state builder, and test to
use Float64 throughout — no conversion needed and aligns with the SDK.

+++ internal/resources/computer/types.go
@@ -16,7 +16,7 @@ type ComputerModel struct {
 	OSPatch                 types.Int64  `tfsdk:"os_patch"`
 	Arch                    types.String `tfsdk:"arch"`
 	CertID                  types.String `tfsdk:"cert_id"`
-	MemorySize              types.Int64  `tfsdk:"memory_size"`
+	MemorySize              types.Float64 `tfsdk:"memory_size"`
 	OSString                types.String `tfsdk:"os_string"`
 	KernelVersion           types.String `tfsdk:"kernel_version"`
 	InstallType             types.String `tfsdk:"install_type"`
